% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.R
\name{nest_into_tree}
\alias{nest_into_tree}
\title{Convert a tibble to a tree of nested tibbles}
\usage{
nest_into_tree(df, name)
}
\arguments{
\item{df}{a dataframe to be converted}

\item{name}{the name to be assigned to the root node}
}
\value{
A tibble containing a sequence of nested tibbles.
 At each level of nesting is a tibble with three columns;
column `name`` contains the levels of the column of the input dataframe at that
level, column `n` contains the number of rows at each subsequent level of nesting,
and column `children` contains the next layer of nested tibbles.
}
\description{
nest_into_tree iterative nests the tibble, forming a tree structure.
The algorithm moves from the left most column, calling `group_by` and `nest`.
Such that the unique values/levels of the left-most column is at the level of
the tree right below the root node, and the levels of the right-most column
form root nodes.
}
